Avance ventas sin ticket

+++ app/Http/Controllers/IngresoProductoController.php
@@ -349,13 +349,12 @@ class IngresoProductoController extends Controller
 
             $producto_barras = $ingreso_producto->producto_barras;
             foreach ($producto_barras as $pb) {
-                if (!$pb->salida_id && !$pb->venta_detalle_id && !$pb->distribucion_detalle_id) {
+                if (!$pb->salida_id && !$pb->venta_id && !$pb->venta_detalle_id && !$pb->distribucion_detalle_id) {
                     $pb->delete();
                 } else {
                     throw new Exception("No es posible eliminar el registro debido a que uno o mas registros del mismo fueron utilizados");
                 }
             }
-
 
             $datos_original = HistorialAccion::getDetalleRegistro($ingreso_producto, "ingreso_productos");
             $ingreso_producto->delete();
